Put utilities needed for testing under tests/
